+++ src/components/Cocktail/CocktailDetails.vue
@@ -9,6 +9,7 @@
             <div class="tag-container" style="margin-bottom: 20px;" v-if="cocktail.tags.length > 0">
                 <RouterLink :to="{name: 'cocktails', query: {'refinementList[tags][0]': tag}}" v-for="tag in cocktail.tags" class="tag tag--background" style="background-color: #BFD3DF;">{{ tag }}</RouterLink>
             </div>
+            <Rating :rating="cocktail.average_rating" type="cocktail" :id="cocktail.id"></Rating>
             <div class="cocktail-details-box__description">
                 <div v-html="parsedDescription"></div>
             </div>
@@ -118,6 +119,7 @@ import Auth from '@/Auth';
 import Unitz from 'unitz'
 import OverlayLoader from '@/components/OverlayLoader.vue'
 import Dropdown from '@/components/Dropdown.vue';
+import Rating from '@/components/Rating.vue';
 
 export default {
     data: () => ({
@@ -130,7 +132,8 @@ export default {
     }),
     components: {
         OverlayLoader,
-        Dropdown
+        Dropdown,
+        Rating
     },
     computed: {
         parsedInstructions() {